backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
httpx>=0.27.0
jieba>=0.42.1
openai>=1.50.0
cryptography>=43.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
sse-starlette>=2.0.0
snownlp>=0.12.3
